refactor: decoupling migration from db
