+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
     "util/fee-estimator",
     "script/data-loader",
     "db",
+    "db-migration",
     "resource",
     "pow",
     "network",
