+++ db-migration/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ckb-db-migration"
+version = "0.34.0-pre"
+license = "MIT"
+authors = ["Nervos Core Dev <dev@nervos.org>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ckb-db = { path = "../db" }
+ckb-logger = { path = "../util/logger" }
+ckb-error = { path = "../error" }
+
+[dev-dependencies]
+tempfile = "3.0"
+ckb-app-config = { path = "../util/app-config" }
