+++ indexer/Cargo.toml
@@ -14,6 +14,7 @@ ckb-jsonrpc-types = { path = "../util/jsonrpc-types" }
 ckb-logger = { path = "../util/logger" }
 ckb-util = { path = "../util" }
 ckb-app-config = { path = "../util/app-config" }
+ckb-db-migration = { path = "../db-migration" }
 
 [dev-dependencies]
 tempfile = "3.0"
