+++ indexer/src/migrations.rs
@@ -1,5 +1,6 @@
 use crate::types::LockHashIndex;
-use ckb_db::{Col, Migration, Result, RocksDB};
+use ckb_db::{Col, Result, RocksDB};
+use ckb_db_migration::Migration;
 use ckb_shared::shared::Shared;
 use ckb_store::ChainStore;
 
@@ -19,7 +20,7 @@ impl AddFieldsToLiveCell {
 }
 
 impl Migration for AddFieldsToLiveCell {
-    fn migrate(&self, db: &RocksDB) -> Result<()> {
+    fn migrate(&self, db: RocksDB) -> Result<RocksDB> {
         const COLUMN_LOCK_HASH_LIVE_CELL: Col = "1";
 
         let snapshot = self.shared.snapshot();
@@ -52,7 +53,8 @@ impl Migration for AddFieldsToLiveCell {
             Ok(())
         };
         db.traverse(COLUMN_LOCK_HASH_LIVE_CELL, migration)?;
-        txn.commit()
+        txn.commit()?;
+        Ok(db)
     }
 
     fn version(&self) -> &str {
