+++ shared/Cargo.toml
@@ -18,3 +18,4 @@ ckb-tx-pool = { path = "../tx-pool" }
 ckb-verification = { path = "../verification" }
 ckb-notify = { path = "../notify" }
 ckb-app-config = { path = "../util/app-config" }
+ckb-db-migration = { path = "../db-migration" }
